backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
langchain-anthropic
langgraph
langchain-core
httpx
tavily-python
yt-dlp
pytest
pytest-asyncio
